Migrate Android and OHOS ports to the delegate API
