+++ ports/servoshell/egl/mod.rs
@@ -10,5 +10,5 @@ mod ohos;
 
 mod log;
 
+mod app_state;
 mod host_trait;
-mod servo_glue;
